Handelling issue with max path length reached.

+++ ripple_unzipple/ripple_unzipple.py
@@ -32,10 +32,9 @@ Usage:
 import os
 import sys
 import time
-import shutil
 from datetime import datetime
-from zipfile import ZipFile
-from py7zr import SevenZipFile
+from zipfile import ZipFile, BadZipFile
+from py7zr import SevenZipFile, Bad7zFile
 from distutils.dir_util import copy_tree
 
 
@@ -100,23 +99,23 @@ def ripple_unzip(input_path, output_path, log_path = ''):
     try:
         # Check if the provided path exists
         if not os.path.exists(input_path):
-            raise ValueError(f"The specified path does not exist: {input_path}")
-        
+            raise ValueError(f"ValueError: The specified path ({input_path}) does not exist")
+
         # Handle different input extensions
         if os.path.isdir(input_path):
             # First copy the directory to the new location 
             copy_tree(input_path, output_path)
-            recursive_unzip(input_path, output_path)
+            recursive_unzip(input_path, output_path, log_path)
         
         elif input_path.endswith((".zip", ".7z")):
             os.makedirs(output_path, exist_ok=True)
 
             with ZipFile(input_path, mode='r') if input_path.endswith(".zip") else SevenZipFile(input_path, mode='r') as archive_ref:
                 archive_ref.extractall(output_path)
-                recursive_unzip(output_path, output_path)
+                recursive_unzip(output_path, output_path, log_path)
         
         else:
-            raise ValueError("Unsupported input type. Please provide a directory or a compressed file.")
+            raise ValueError("ValueError: Unsupported input type. Please provide a directory or a compressed file.")
         
     except ValueError as error:
         logging(log_path, Colors.ERROR, str(error))
@@ -125,13 +124,14 @@ def ripple_unzip(input_path, output_path, log_path = ''):
         logging(log_path, Colors.ERROR, str(error))
         raise Exception(error)  
     
-def recursive_unzip(input_path, output_path):   
+def recursive_unzip(input_path, output_path, log_path = ''):   
     """
     Recursively unzip .zip and .7z files in the input_path to the output_path.
 
     Args:
         input_path (str): Path to the input directory or compressed file.
         output_path (str): Path to the output directory.
+        log_path (str, optional): Path to the log file. Defaults to ''.
         
     Return:
         None.
@@ -150,17 +150,28 @@ def recursive_unzip(input_path, output_path):
 
             file_to_remove = ''
             if file.endswith((".zip", ".7z")):
-                with ZipFile(file_path, mode='r') if file.endswith(".zip") else SevenZipFile(file_path, mode='r') as archive_ref:
-                    archive_ref.extractall(extract_path)
-                    
-                    # Recursively call the function to check every file in the directory tree
-                    recursive_unzip(extract_path, extract_path)
-                    
-                    # Flag the compressed file to be removed
-                    file_to_remove = extract_path + '.zip' if file.endswith(".zip") else extract_path + '.7z'
+                try:    
+                    with ZipFile(file_path, mode='r') if file.endswith(".zip") else SevenZipFile(file_path, mode='r') as archive_ref:
+                        archive_ref.extractall(extract_path)
+
+                        # Recursively call the function to check every file in the directory tree
+                        recursive_unzip(extract_path, extract_path, log_path)
+
+                        # Flag the compressed file to be removed
+                        file_to_remove = extract_path + '.zip' if file.endswith(".zip") else extract_path + '.7z'
+                except FileNotFoundError as error:
+                    error_message = f"FileNotFoundError: {error.strerror} in ({file_path})\n\nA common cause for this issue may be that the MAX_PATH_LENGTH for your machine's directory is surpassed. The compressed directory will be placed in the folder for you to extract manually. Please read the Configuration section in the README to resolve this issue."
+                    logging(log_path, Colors.ERROR, error_message)
+                    continue
+                except (BadZipFile, Bad7zFile) as error:
+                    error_message = f"BadZipFile or Bad7ZFile: {error.strerror} with ({file_path})\n\nContinuing the tool and placing file for manual extracting."
+                    logging(log_path, Colors.ERROR, error_message)
+                    continue
 
                 # Remove the original compressed file from the new output folder
-                os.remove(file_to_remove)
+                if os.path.exists(file_to_remove):
+                    os.remove(file_to_remove)
+
 
             
 #========================================================
@@ -172,21 +183,22 @@ def main():
     start_time = time.time()
     print(f'\n{Colors.INFO}Tool is starting...{Colors.END}')
     
+    log_path = ''
     try:
         # Get the path from the command-line argument
         if len(sys.argv) not in [3, 4]:
-            raise ValueError("Usage: python ripple_unzipple.py input_path output_path [log_path]")
+            raise ValueError("Usage: python path/to/ripple_unzipple.py input_path output_path [log_path]")
 
         # Call the recursive_unzip function with the provided path
         log_path = ''
         if len(sys.argv) == 4:
-            ripple_unzip(sys.argv[1], sys.argv[2], sys.argv[3])
             log_path = sys.argv[3]
 
         # Call the main logic function for starting the recursive unzipping
-        ripple_unzip(sys.argv[1], sys.argv[2])
+        ripple_unzip(sys.argv[1], sys.argv[2], log_path)
+
     except Exception as error:
-        # logging(log_path, Colors.ERROR, error)
+        logging(log_path, Colors.ERROR, error)
         exit(1)
         
     # Get the end time of the script and calculate the elapsed time
